Extract `ChatPageModel` and `NavigationModel` to their own files.
