+++ client/src/main/kotlin/io/spine/examples/chatspn/desktop/chat/ChatPage.kt
@@ -353,234 +353,6 @@ private fun EditMessagePanel(model: ChatPageModel) {
 }
 
 /**
- * UI Model for the [ChatPage].
- *
- * UI Model is a layer between `@Composable` functions and client.
- */
-private class ChatPageModel(
-    val client: DesktopClient,
-    var chatData: ChatData,
-    val openChatInfo: (chat: ChatId) -> Unit,
-    val openUserProfile: (user: UserId) -> Unit
-) {
-    private val messagesState: MutableMessagesState = MutableStateFlow(listOf())
-    val messageInputFieldState: MessageInputFieldState = MessageInputFieldState()
-    val chatDeletionModalState: MutableState<Boolean> = mutableStateOf(false)
-
-    /**
-     * Reads messages in the chat and subscribes to their updates.
-     */
-    fun observeMessages() {
-        messageInputFieldState.clear()
-        messagesState.value = client.readMessages(chatData.id).toMessageDataList(client)
-        client.stopObservingMessages()
-        client.observeMessages(
-            chatData.id,
-            { messageView -> updateMessagesState(messageView) },
-            { messageDeleted -> updateMessagesState(messageDeleted) })
-    }
-
-    /**
-     * Returns the state of messages in the chat.
-     */
-    fun messages(): MessagesState {
-        return messagesState
-    }
-
-    /**
-     * Updates the state of chat messages by adding a new message,
-     * or editing existing one if message ID matches.
-     *
-     * @param messageView message to update the state
-     */
-    private fun updateMessagesState(messageView: MessageView) {
-        val message = messageView.toMessageData(client)
-        val chatMessages = messagesState.value
-        if (chatMessages.findMessage(message.id) != null) {
-            val newChatMessages = chatMessages.replaceMessage(message)
-            messagesState.value = newChatMessages
-        } else {
-            messagesState.value = chatMessages + message
-        }
-    }
-
-    /**
-     * Updates the state of chat messages by removing a message.
-     *
-     * @param messageDeleted event about message deletion
-     */
-    private fun updateMessagesState(messageDeleted: MessageMarkedAsDeleted) {
-        val chatMessages = messagesState.value
-        val message = chatMessages.findMessage(messageDeleted.id)
-        if (message != null) {
-            val messageIndex = chatMessages.indexOf(message)
-            val newChatMessages = chatMessages.remove(messageIndex)
-            messagesState.value = newChatMessages
-        }
-    }
-
-    /**
-     * Sends a message to the selected chat.
-     *
-     * @param content message text content
-     */
-    fun sendMessage(content: String) {
-        client.sendMessage(chatData.id, content)
-    }
-
-    /**
-     * Opens the message editing panel.
-     *
-     * @param message message data to start editing
-     */
-    fun startMessageEditing(message: MessageData) {
-        messageInputFieldState.isEditingState.value = true
-        messageInputFieldState.editingMessage.value = message
-        messageInputFieldState.inputText.value = message.content
-    }
-
-    /**
-     * Edits message in the selected chat.
-     *
-     * @param message ID of the message to edit
-     * @param newContent new text content for the message
-     */
-    fun editMessage(message: MessageId, newContent: String) {
-        client.editMessage(chatData.id, message, newContent)
-    }
-
-    /**
-     * Deletes the chat.
-     *
-     * @param chat ID of the chat to delete
-     */
-    fun deleteChat(chat: ChatId) {
-        client.deleteChat(chat)
-    }
-
-    /**
-     * State of the message input field.
-     */
-    class MessageInputFieldState {
-        val inputText: MutableState<String> = mutableStateOf("")
-        val isEditingState: MutableState<Boolean> = mutableStateOf(false)
-        val editingMessage: MutableState<MessageData?> = mutableStateOf(null)
-
-        /**
-         * Clears the state.
-         */
-        fun clear() {
-            inputText.value = ""
-            isEditingState.value = false
-            editingMessage.value = null
-        }
-    }
-}
-
-/**
- * Data for the single message view.
- */
-public data class MessageData(
-    val id: MessageId,
-    val sender: UserProfile,
-    val content: String,
-    val whenPosted: Timestamp
-)
-
-/**
- * Finds message in the list by ID.
- *
- * @param id ID of the message to find
- * @return found message or `null` if message is not found
- */
-private fun MessageList.findMessage(id: MessageId): MessageData? {
-    val message = this
-        .stream()
-        .filter { message -> message.id.equals(id) }
-        .collect(Collectors.toList())
-    if (message.isEmpty()) {
-        return null
-    }
-    return message[0]
-}
-
-/**
- * Returns the new list with the replaced message.
- *
- * @param newMessage message to replace
- */
-private fun MessageList.replaceMessage(newMessage: MessageData): MessageList {
-    val oldMessage = this.findMessage(newMessage.id)
-    val messageIndex = this.indexOf(oldMessage)
-    val leftPart = this.subList(0, messageIndex)
-    val rightPart = this.subList(messageIndex + 1, this.size)
-    return leftPart + newMessage + rightPart
-}
-
-/**
- * Creates the `MessageData` from the `MessageView`.
- *
- * @param client desktop client to find user profiles
- */
-private fun MessageView.toMessageData(client: DesktopClient): MessageData {
-    return MessageData(
-        this.id,
-        client.findUser(this.user)!!,
-        this.content,
-        this.whenPosted
-    )
-}
-
-/**
- * Creates the `MessageData` list. from the `MessageView` list.
- *
- * @param client desktop client to find user profiles
- */
-private fun List<MessageView>.toMessageDataList(client: DesktopClient): MessageList {
-    val users = mutableMapOf<UserId, UserProfile>();
-    val messages = this.stream().map { message ->
-        val user: UserProfile
-        if (users.containsKey(message.user)) {
-            user = users[message.user]!!
-        } else {
-            user = client.findUser(message.user)!!
-            users[message.user] = user
-        }
-        MessageData(
-            message.id,
-            user,
-            message.content,
-            message.whenPosted
-        )
-    }.collect(Collectors.toList())
-    return messages
-}
-
-/**
- * List of `MessageData`.
- */
-public typealias MessageList = List<MessageData>
-
-/**
- * Mutable state of messages in the chat.
- */
-private typealias MutableMessagesState = MutableStateFlow<MessageList>
-
-/**
- * Immutable state of messages in the chat.
- */
-public typealias MessagesState = StateFlow<MessageList>
-
-/**
- * Returns the new list without an element on provided index.
- *
- * @param index index of the element to remove
- */
-private fun <T> List<T>.remove(index: Int): List<T> {
-    return this.subList(0, index) + this.subList(index + 1, this.size)
-}
-
-/**
  * Defines whether the message at the provided index is the first in the message sequence
  * from a particular user in the last 10 minutes.
  */
